README file updated

+++ config/migrations.sql
@@ -21,7 +21,7 @@ USE `orizon_offers`;
 --
 
 CREATE TABLE `countries` (
-  `country_id` tinyint(3) NOT NULL,
+  `country_id` int(3) NOT NULL,
   `name` varchar(35) NOT NULL
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8 COLLATE=utf8_bin;
 
@@ -53,9 +53,9 @@ INSERT INTO `countries` (`country_id`, `name`) VALUES
 
 CREATE TABLE `travels` (
   `travel_id` int(11) NOT NULL,
-  `departure_id` tinyint(3) NOT NULL,
-  `destination_id` tinyint(3) NOT NULL,
-  `available_places` tinyint(3) NOT NULL
+  `departure_id` int(3) NOT NULL,
+  `destination_id` int(3) NOT NULL,
+  `available_places` int(3) NOT NULL
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8 COLLATE=utf8_bin;
 
 --
@@ -105,7 +105,7 @@ ALTER TABLE `travels`
 -- AUTO_INCREMENT per la tabella `countries`
 --
 ALTER TABLE `countries`
-  MODIFY `country_id` tinyint(3) NOT NULL AUTO_INCREMENT, AUTO_INCREMENT=15;
+  MODIFY `country_id` int(3) NOT NULL AUTO_INCREMENT, AUTO_INCREMENT=15;
 
 --
 -- AUTO_INCREMENT per la tabella `travels`
